add map markers

+++ app/src/main/java/com/example/mapbox_test_task/model/Properties.kt
@@ -5,9 +5,9 @@ data class Properties (
     val extrude : Boolean,
     val iso_3166_1 : String,
     val underground : Boolean,
-    val height : Int,
+    val height : Float,
     val type : String,
-    val min_height : Int,
+    val min_height : Float,
     val iso_3166_2 : String,
     val tilequery : Tilequery
 )
